+++ src/directives/model/select.js
@@ -58,7 +58,7 @@ function initOptions (expression) {
   var self = this
   function optionUpdateWatcher (value) {
     if (_.isArray(value)) {
-      self.el.innerHTML = value.map(formatOption)
+      self.el.innerHTML = value.map(formatOption).join('')
       if (self._watcher) {
         self.update(self._watcher.value)
       }
@@ -87,9 +87,15 @@ function initOptions (expression) {
 
 function formatOption (op) {
   if (typeof op !== 'string') {
-    return '<optgroup label="' + op.label + '">' +
-      op.options.map(formatOption) +
+    if (op.value) {
+      return '<option value="' + op.value + '">' +
+        op.label +
+      '</option>'
+    } else if (op.options) {
+      return '<optgroup label="' + op.label + '">' +
+        op.options.map(formatOption).join('') +
       '</optgroup>'
+    }
   } else {
     return '<option>' + op + '</option>'
   }
@@ -102,9 +108,8 @@ function formatOption (op) {
 function checkInitialValue () {
   var initValue
   var options = this.el.options
-  var i = options.length
-  while (i--) {
-    if (options[i].selected) {
+  for (var i = 0, l = options.length; i < l; i++) {
+    if (options[i].hasAttribute('selected')) {
       if (this.multiple) {
         (initValue || (initValue = []))
           .push(options[i].value)